backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
openai==1.57.0
qdrant-client>=1.8.0
sqlalchemy>=1.4.52
pydantic>=2.5.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.0
pytest>=7.4.3
httpx==0.27.2
PyYAML>=6.0
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
requests>=2.31.0
mangum>=0.17.0